Fix the caption spacing
This changes the line spacing in long figure and table captions to
single spaced per Chapter 10 of the April 2014 thesis manual.  The label
is still bold.  I also reduced the size of the caption to add
distinction from the main text.  This update now requires the `caption`
and `setspace` packages.